// src/screens/fintechApp/payments/interface.tsx
export interface PaymentLinkListInterface {
    id: string;
    customerId: string;
    date: Date;
    amount: number,
    currency: string;
    network: string;
    type: string;
    walletAddress: string;
    state: string;
    merchantName: string;
    invoiceNo:string;
}